Remove ProxyBackendConfiguration
